refactor: move extract_tar_zst into a dedicated archive module at the crate root
